Move `WaitGroup` to package `utils`
