+++ cmd/utils/wait_group.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package cmd
+package utils
 
 import "sync"
